Implemented CI for main branch
Now, when the main branch is updated -- or a PR to the main branch is updated -- a Github Actions workflow is triggered. At this time it builds the two Kotlin/KVM microservices, runs the tests, and generates Docker images.

What's left to do: build the UI app, generate a Docker image, and post all the images to a registry.